Port of 2.0-beta bug fix by Mark Tsuchida.  Fix large memory consumption of multiposition file writing
The main culprit was the cache of 10 image-sized ByteBuffer instances
per writer, which is per position if saving to separate file per
position.

For example, if a single image is 8 MiB and there are 900 positions,
this would consume 70 GiB, even though no benefit is expected at the
data rate of most multiposition acquisitions. After this fix, only 24
MiB is consumed.

Changed MultipageTiffWriter to use a single, small cache of buffers
shared across all writers.

Also ensures only direct buffers are recycled, and that the most
recently used buffer is recycled first.

Not optimized for the (normally unexpected) case where multiple files
containing different-sized images are being written, but the overhead in
that case should be insignificant.