Studio: Fix live display of multi-channel cameras.
Previously, the code looked at the sequence number and only displayed an
image if the new sequence number was higher than the previous one. Since
multi-channel cameras have different sequences for each channel (or at
least, this is a possible implementation), the assumption of a single
sequence number was false.  The code now maintains knowledge of the
sequence number of each individual channel.

Also, the name rather than the number of each cameraChannel is used now by
the UI.

Cleanup of various warnings..