Studio: Fixed bug in Property editor
This bug manifested itself when refreshing when not all checkboxes were checked.
Afterwards, checking an additional checkbox would not result in additional
properties appearing.

Also took the opportunity to refactor the property code a bit.  There is
still quite a bit of chaos there, and would benefit from more cleanup
(for which I currently do not have time)