Windows installer: completely delet micro-manager dir in plugins.
It becomes extremely unwieldy to avoid mutiple version of the same jar,
and not deleting this directory will lead to name clashes that prevent
the application from starting up.  I doubt that anyone will put
anything into this directory themselves...

Also force the installer to always show the installation directory
(and ask for its location in the beginning).  Not seeing that dialog
makes me very nervous (not sure why it did not show it, but better
be safe than sorry in the future).