Studio: Re-added json source code, reverting previous commit.
Using any of the maven org.json jars results in the run-time error:
JSONArray initial value should be a string or collection or array.
These errors are coming from the Clojure acquisition engine code and
appear to be caused by the way the Clojure code creates JSONArrays.
This is all too mysterious to trouble shoot, and rather wait until the
Clojure code can be removed.