Andor: It is no longer necessary to run live before snap starts working.
The driver returns an error code for the IsTriggerModeavailable function in UpdateSnapTriggerMode. 
Therefore, now test whether or not this camera supports software trigger mode by checking the 
MM property containing all trigger modes.  I do not know why the driver sometimes says that 
software trigger is not available, but this work-around does the trick.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17180 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd