Studio: Viewer now displays correct channel name in composite mode.
Previously, the channel name would be frozen on the last used name,
but the user could still move the slider.  It would have been nice to
show the names of all channels in Composite mode and disable the channels
slider, but that would need a lot of tangled up code.
This commit always shows the correct channel name for the channel
slider position, whether in composite mode or not.