Studio: Try to avoid multiple live mode suspensions when using QUickAccess channels.
Bug reported by Harris Morrison.  When using channels in the QuickAccess panel while live mode is running,
mulitple shutter openings and closings occur, and every channel change induces more of these.
This code runs through refreshGUI(), which causes lots of hardware interactions.  I chaged this to
refreshGUIFromChache() (the configGroupPad in the main window uses the non-API updateGUI(false) function).

I also changed the function MainFrame.configureBinningComboForCamera, so that the action listeners
are removed before removing the item so that the actionlisteneres are not executed (as they would
also stop and restart live mode).

Will need user feedback to see if this fixed everything.