Point and Shoot Analysis: New bleach tracking mode.
Added ability to track a bleach point by first determining its relation
to the centroid of the particle it is in, and then locating it in newer
frames based on this vector.  This mode assumes that the particle does
not change shape or rotates.  The UI lets the user activate this mode,
and adds input for the number of frames that will be used to establish
the centroid-blach spot offset.

Also makes sure that the fits use the same color scheme as the original
data making it easier to compare where everything is coming from.