ZProjector: work around bug in Multi-Page Tiff saving code.
Projecting a large data set that misses data in the last time point.
This caused the problem descibed in issue #716.  By changing the order
of Coords, such that the time axis is worked on last, results in
saving the data correctly.  I am not absolutely sure that other
orders do not work (will test), and clearly, the current solution
only works when missing data at the end of the time axis, not of
other axes.