Studio: Handle DPNewImageEvent only if we are not behind handling the sequence buffer.
The intent of the previous code is to only handle new images if
the display can keep up.  However, this code can take 10-20ms to run,
and since it (was) called synchronously the datastore, it could interfere
with data saving.  Even though that eventbus is now asynchronous, this
code may spend extensive resources when the focus should be on saving
data.  So, this commit check if and acquisition is running and if there
are more than a certain number of images in the sequence buffer, and
only continues when things are calm. This made it possible again to
save acquisitions of 8MB images at 100fps.